server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
asyncpg>=0.30.0
psycopg[binary]>=3.2.0
pydantic-settings>=2.6.0
email-validator>=2.2.0
python-multipart>=0.0.12
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
supabase>=2.8.0
httpx>=0.28.0
pypdf>=5.1.0
python-docx>=1.1.2
Pillow>=11.0.0
pytesseract>=0.3.13
pypdfium2>=4.30.0
